requirements.txt
numpy
Pillow
tqdm
pygame